requirements.txt
requests
tqdm
pytest-playwright
